wcore -43 -  Renaming the contract root object, since it isn't specific to s_roots.
